# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Dokumentasi BKPSDM Katingan",
        "description": "Ini adalah dokumentasi API yang digunakan untuk website BKPSDM Katingan.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/agendas/{slug}/download": {
            "get": {
                "tags": [
                    "Agenda"
                ],
                "summary": "Mengunduh file agenda",
                "operationId": "downloadAgendaFile",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "contoh-agenda-unik"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File berhasil diunduh",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "Data atau file tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "File tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token yang didapat dari endpoint /api/login. Contoh: 1|aBcDeFgHiJkLmNoPqRsTuVwXyZ",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}